requirements.txt
torch==1.12.1+cu113
torchvision==0.13.1+cu113
pandas
scikit-learn
PyYAML
albumentations
tqdm
segmentation-models-pytorch
timm==0.6.12
opencv-python
monai
ttach  # test time augmentation
torchmetrics
transformers
pytorch_lightning
pickle5
matplotlib

nilearn
monai-weekly[nibabel, tqdm]